Process metrics into influxdb format.
This is done on a per-collection-group basis.  Since influxdb dones't
care about regular collection intervals, we may as well re-use the
collection_group logic, and send to influxdb at the same time.

Since influxdb can take arbitrary-sized packets (theoretically up to the
64k--the maximum size of a UDP packet), we attempt to bundle multiple
metrics in a collection group into a single packet.  This is still a
work in progress, but that's the idea...